+ Top Package Descriptions
In the Install Packages menu, rows in the Top Packages sections now show their
description (if found), matching the behavior of rows in the search results.

To accomplish this, Top Package list items now use instances of SearchResultRow
instead of their own unique View type, allowing this code be deduplicated.