~ Lazily Get Package Size
Previously, the size of each installed package was eagerly calculated at launch.
However, this is expensive, and can be unnecessary in many cases.
(This property's main use is for sorting the sidebar list by package size,
so if another sort mode is selected, the compute is wasted.)

Package sizes are now calculated lazily. To support this, package models
are initialized with their URL instead of the size value proper.
The size is cached to prevent it from being re-computed on every access.