^ Zombie process when installation cancelled
Previously, cancelling an installation would not terminate the child process.
So, if Brew were to hang (eg due to waiting for input), it would remain as a
zombie until manually killed via Activity Monitor or the like.

Due to Brew's locking mechanisms, this also means that no other process would
be able to access the package in question.

To prevent this, a `shell` overload returning the spawned process has now been
added, allowing InstallationProgressTracker to terminate it upon cancellation.